+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 # Unreleased
 * perf: add boringssl for p256Verify precompile [#275](https://github.com/hyperledger/besu-native/pull/275)
+* perf: add higher performing ecrecover-specific entrypoints for secp256k1/r1 [#277](https://github.com/hyperledger/besu-native/pull/277)
 
 # 1.3.1
 * perf: use latest gnark-crypto 0.17 [#262](https://github.com/hyperledger/besu-native/pull/262)